implementando conexion a entre los microservicios

+++ microservice.role/src/main/java/com/microservices/role/repositories/IRolREpository.java
@@ -1,13 +1,13 @@
 package com.microservices.role.repositories;
 
-import org.apache.catalina.Role;
+import com.microservices.role.entities.Role;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 import java.util.List;
 
 @Repository
-public interface IRolREpository extends JpaRepository<Role, Integer> {
+public interface IRolREpository extends JpaRepository<Role, Long> {
